Re designed, query patern for user service
