+++ src/lib/objectToProps.ts
@@ -1,0 +1,7 @@
+const objectToProps = <T>(values: { [key: string]: T }) => ({
+  props: {
+    ...values
+  }
+})
+
+export default objectToProps
