--- src/lib/query/helpers.ts
@@ -1,7 +1,0 @@
-export const objectAsProps = <T>(values: T) => ({
-  props: {
-    ...values
-  }
-})
-
-export default {}
