--- src/lib/query/index.ts
@@ -1,5 +1,0 @@
-import query from "./query"
-
-export { objectAsProps } from "./helpers"
-
-export default query
