--- src/lib/query/queries/index.ts
@@ -1,3 +1,0 @@
-export { getAllUsers } from "./users"
-
-export default {}
