--- src/lib/query/transforms/index.ts
@@ -1,2 +1,0 @@
-export { getAllUsers } from "./users"
-export type { AllUsers } from "./users"
