+++ src/lib/useCases/index.ts
@@ -1,0 +1,2 @@
+export { getAllUsers } from "./getAllUsers"
+export type { Users } from "./getAllUsers"
